requirements.txt
requests
python-dotenv
numpy
transformers
torch